Move color codes into one file
