+++ aristo/info.py
@@ -1,16 +1,8 @@
 import os, io, platform, sys, socket
 from time import sleep
 from urllib.request import urlopen
-
-W  = '\033[0m'  # white (normal)
-R  = '\033[31m' # red
-G  = '\033[32m' # green
-O  = '\033[33m' # orange
-B  = '\033[34m' # blue
-P  = '\033[35m' # purple
-C  = '\033[36m' # cyan
-GR = '\033[37m' # gray
-T  = '\033[93m' # tan 
+import color
+from color import *
 
 mac_address = os.popen("cat /sys/class/net/eth0/address").read()
 s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
